+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Ribbon V2.11.1, RibbonTools V1.3.7+
+
+#### Changed (RibbonTools)
+
+- Added functions (and small changes) to AlphaBitmap, NativeMethods
+
 ### Ribbon V2.11.1, RibbonTools V1.3.7
 
 #### Changed (RibbonTools)